+++ dev-services/postgres/data/seed.sql
@@ -12,7 +12,7 @@ CREATE TABLE events (
 	event_name VARCHAR(255) NOT NULL,
 	event_description VARCHAR(2000) NOT NULL,
 	event_date DATE NOT NULL,
-	event_time TIMESTAMP NOT NULL
+	event_time TIME NOT NULL
 );
 
 CREATE TABLE rsvps (